# artifacts/karma-ai/src/lib/api-client-react/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type GeminiChatMessageRole = typeof GeminiChatMessageRole[keyof typeof GeminiChatMessageRole];


export const GeminiChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface GeminiChatMessage {
  role: GeminiChatMessageRole;
  /** @minLength 1 */
  content: string;
}

export type GeminiChatInputImagesItem = {
  mimeType?: string;
  data?: string;
};

export interface GeminiChatInput {
  /**
     * @minLength 1
     * @maxLength 12000
     */
  message: string;
  /** @maxItems 20 */
  history?: GeminiChatMessage[];
  /**
     * @maxLength 50000
     * @nullable
     */
  materialText?: string | null;
  /**
     * @maxLength 200
     * @nullable
     */
  materialName?: string | null;
  /** @maxItems 1 */
  images?: GeminiChatInputImagesItem[];
}

export interface GeminiChatResponse {
  message: string;
  model: string;
  groundedInMaterial: boolean;
}

export interface LearningRecommendationsQuery {
  /**
     * @minLength 2
     * @maxLength 200
     */
  query: string;
  /**
     * @minimum 1
     * @maximum 12
     */
  limit?: number;
}

export type RecommendationSourceId = typeof RecommendationSourceId[keyof typeof RecommendationSourceId];


export const RecommendationSourceId = {
  igot: 'igot',
  nssta: 'nssta',
} as const;

export type RecommendationSourceStatus = typeof RecommendationSourceStatus[keyof typeof RecommendationSourceStatus];


export const RecommendationSourceStatus = {
  available: 'available',
  unavailable: 'unavailable',
} as const;

export interface RecommendationSource {
  id: RecommendationSourceId;
  name: string;
  status: RecommendationSourceStatus;
  endpoint: string;
  /** @nullable */
  message: string | null;
}

export type LearningRecommendationSource = typeof LearningRecommendationSource[keyof typeof LearningRecommendationSource];


export const LearningRecommendationSource = {
  igot: 'igot',
  nssta: 'nssta',
} as const;

export type LearningRecommendationType = typeof LearningRecommendationType[keyof typeof LearningRecommendationType];


export const LearningRecommendationType = {
  course: 'course',
  programme: 'programme',
} as const;

export interface LearningRecommendation {
  id: string;
  source: LearningRecommendationSource;
  sourceName: string;
  title: string;
  summary: string;
  type: LearningRecommendationType;
  /**
     * @minimum 0
     * @maximum 100
     */
  relevance: number;
  destinationUrl: string;
  /** @nullable */
  provider: string | null;
  /** @nullable */
  duration: string | null;
  /** @nullable */
  schedule: string | null;
}

export type LearningRecommendationsResponseCatalogueStatus = typeof LearningRecommendationsResponseCatalogueStatus[keyof typeof LearningRecommendationsResponseCatalogueStatus];


export const LearningRecommendationsResponseCatalogueStatus = {
  available: 'available',
  partial: 'partial',
  unavailable: 'unavailable',
} as const;

export interface LearningRecommendationsResponse {
  query: string;
  catalogueStatus: LearningRecommendationsResponseCatalogueStatus;
  recommendations: LearningRecommendation[];
  sources: RecommendationSource[];
}

export interface ErrorResponse {
  error: string;
}

export type GetLearningRecommendationsParams = {
/**
 * Competency, topic, or learning goal to search for
 * @minLength 2
 * @maxLength 200
 */
query: string;
/**
 * Maximum number of recommendations to return
 * @minimum 1
 * @maximum 12
 */
limit?: number;
};

export interface QuizQuestion {
  id: number;
  question: string;
  options: string[];
  correctIndex: number;
  explanation: string;
}

export interface QuizResponse {
  title: string;
  questions: QuizQuestion[];
}

export interface QuizGenerateResponse {
  quiz: QuizResponse;
}