# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /gemini/chat:
    post:
      operationId: sendGeminiChat
      tags: [gemini]
      summary: Ask KARMA AI a learning question
      description: Sends a learning question and optional uploaded-material context to the Gemini-powered KARMA AI assistant.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeminiChatInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiChatResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: AI provider error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    GeminiChatMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
          minLength: 1
      required:
        - role
        - content
    GeminiChatInput:
      type: object
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 12000
        history:
          type: array
          maxItems: 20
          items:
            $ref: "#/components/schemas/GeminiChatMessage"
        materialText:
          type: ["string", "null"]
          maxLength: 50000
        materialName:
          type: ["string", "null"]
          maxLength: 200
        images:
          type: array
          maxItems: 1
          items:
            type: object
            properties:
              mimeType:
                type: string
              data:
                type: string
            required: []
        provider:
          type: string
        model:
          type: string
      required:
        - message
    GeminiChatResponse:
      type: object
      properties:
        message:
          type: string
        model:
          type: string
        groundedInMaterial:
          type: boolean
        provider:
          type: string
      required:
        - message
        - model
        - groundedInMaterial
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...