# lib/api-zod/src/generated/types/geminiChatInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GeminiChatInputImagesItem } from './geminiChatInputImagesItem';
import type { GeminiChatMessage } from './geminiChatMessage';

export interface GeminiChatInput {
  /**
     * @minLength 1
     * @maxLength 12000
     */
  message: string;
  /** @maxItems 20 */
  history?: GeminiChatMessage[];
  /**
     * @maxLength 50000
     * @nullable
     */
  materialText?: string | null;
  /**
     * @maxLength 200
     * @nullable
     */
  materialName?: string | null;
  /** @maxItems 1 */
  images?: GeminiChatInputImagesItem[];
  provider?: string;
  model?: string;
}